last commit

+++ exercise_session_05/init.cpp
@@ -95,7 +95,7 @@ double boundary(double x, double y, int rhs_function){
  *                   The "interior" entries of f store the source term of Poisson eq.
  */
 void init_f(params p, double **f){
-    double delta = 1 / (p.nx - 1);
+    double delta = 1.0 / (double)(p.nx - 1);
     for (int i = 0; i<p.nx; i++){
         for (int j = 0; j<p.ny; j++){
             if ((i == 0) || (j == 0) || (i == p.nx-1) || (j == p.ny-1)) {
